Also pass Container to ManifestClassPathUtils instead of only Entry
- ManifestClassPathUtils methods take an Entry object that is adapted
from a Container object.  Before this PR the Entry was adapted back to a
Container object just to call getEntry on it.  Instead of making a whole
new Container object, pass in the original Container that was adapted to
the Entry.
